:sparkles: chore: create private route and protect routes
