+++ web/src/App.tsx
@@ -1,7 +1,7 @@
 import React from 'react';
 import { ThemeProvider } from 'styled-components';
 
-import Routes from './routes';
+import Routes from './routes/routes';
 
 import theme from './styles/theme';
 import GlobalStyle from './styles/global';
